+++ DSCResources/Outlook2013/Outlook2013.schema.psm1
@@ -97,4 +97,7 @@ Configuration Outlook2013
 
     Import-DscResource -ModuleName xPSDesiredStateConfiguration -ModuleVersion 8.3.0.0
     . "$resourcePath\windows.xRegistry.ps1"
+
+    Import-DscResource -ModuleName xPSDesiredStateConfiguration -ModuleVersion 8.3.0.0 #jjs continue to work the applicable module required.
+    . "$resourcePath\windows.cAdministratorTemplate.ps1"
 }