+++ Module/Convert.Stig/Data.ps1
@@ -14,7 +14,7 @@ data DscResource
         MimeTypeRule                    = xIisMimeTypeMapping
         ManualRule                      = None
         ProcessMitigationRule           = ProcessMitigation
-        RegistryRule                    = xRegistry
+        RegistryRule                    = xRegistry, cAdministratorTemplate #jjs
         SecurityOptionRule              = SecurityOption
         ServiceRule                     = xService
         SqlScriptQueryRule              = SqlScriptQuery
@@ -26,3 +26,4 @@ data DscResource
         WmiRule                         = Script
 '@
 }
+#jjs need to address RegistryRule between xRegistry and cAdministratorTemplate